phish-sim/inference/requirements.txt
# SPDX-License-Identifier: MIT
# Copyright (c) 2024 Phish-Sim Project

# Web framework and API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
starlette==0.27.0

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Redis for caching and session management
redis==5.0.1
aioredis==2.0.1
redis-py-cluster==2.1.3

# Message queuing and task processing
celery==5.3.4
kombu==5.3.4
rq==1.15.1
dramatiq==1.15.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# HTTP client and networking
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Real-time processing
asyncio-mqtt==0.16.1
aiokafka==0.8.11
pika==1.3.2

# Monitoring and metrics
prometheus-client==0.19.0
grafana-api==1.0.3
psutil==5.9.6
py-cpuinfo==9.0.0

# Logging and debugging
structlog==23.2.0
loguru==0.7.2
sentry-sdk==1.38.0

# Configuration and environment
python-dotenv==1.0.0
pydantic-settings==2.1.0
pyyaml==6.0.1
toml==0.10.2

# Security and authentication
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Data processing and serialization
orjson==3.9.10
msgpack==1.0.7
pickle5==0.0.12

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
httpx==0.25.2

# Performance and optimization
uvloop==0.19.0
gunicorn==21.2.0
gevent==23.9.1

# Utilities
click==8.1.7
tqdm==4.66.1
rich==13.7.0
typer==0.9.0

# Machine learning integration
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Time and scheduling
croniter==2.0.1
apscheduler==3.10.4